#1596: elm: abstract ElementStats and move to elm
