+++ src/vt/elm/elm_stats.cc
@@ -41,13 +41,12 @@
 //@HEADER
 */
 
-#include "vt/config.h"
-#include "vt/vrt/collection/balance/elm_stats.h"
-#include "vt/timing/timing.h"
+#if !defined INCLUDED_VT_ELM_ELM_STATS_CC
+#define INCLUDED_VT_ELM_ELM_STATS_CC
 
-#include <cassert>
+#include "vt/elm/elm_stats.h"
 
-namespace vt { namespace vrt { namespace collection { namespace balance {
+namespace vt { namespace elm {
 
 void ElementStats::startTime() {
   auto const start_time = timing::Timing::getCurrentTime();
@@ -242,20 +241,6 @@ std::size_t ElementStats::getSubphaseCommPhaseCount() const {
   return subphase_comm_.size();
 }
 
-/*static*/
-void ElementStats::setFocusedSubPhase(VirtualProxyType collection, SubphaseType subphase) {
-  focused_subphase_[collection] = subphase;
-}
+}} /* end namespace vt::elm */
 
-/*static*/
-SubphaseType ElementStats::getFocusedSubPhase(VirtualProxyType collection) {
-  auto i = focused_subphase_.find(collection);
-  if (i != focused_subphase_.end())
-    return i->second;
-  else
-    return no_subphase;
-}
-
-/*static*/ std::unordered_map<VirtualProxyType,SubphaseType> ElementStats::focused_subphase_;
-
-}}}} /* end namespace vt::vrt::collection::balance */
+#endif /*INCLUDED_VT_ELM_ELM_STATS_CC*/
