+++ src/vt/vrt/collection/manager.impl.h
@@ -1377,7 +1377,7 @@ void CollectionManager::insertMetaCollection(
     using MsgType = CollectStatsMsg<ColT>;
     auto const phase = thePhase()->getCurrentPhase();
     CollectionProxyWrapType<ColT> p{bits};
-    p.template broadcastCollective<MsgType,ElementStats::syncNextPhase<ColT>>(
+    p.template broadcastCollective<MsgType,CollectionStats::syncNextPhase<ColT>>(
       phase
     );
   };
